Various fixes, modelbrowser should partially work now

+++ src/vgui2/vgui_controls/Panel.cpp
@@ -3551,7 +3551,7 @@ void Panel::RequestFocus(int direction)
 //-----------------------------------------------------------------------------
 void Panel::OnRequestFocus(VPANEL subFocus, VPANEL defaultPanel)
 {
-	CallParentFunction(new KeyValues("OnRequestFocus", "subFocus", (int)subFocus, "defaultPanel", (int)defaultPanel));
+	CallParentFunction(new KeyValues("OnRequestFocus", "subFocus", subFocus, "defaultPanel", defaultPanel));
 }
 
 //-----------------------------------------------------------------------------
